requirements.txt
# pip install -r requirements.txt

# Core Django REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication Packages
djangorestframework-simplejwt==5.3.1
dj-rest-auth[with_social]==5.1.2
django-allauth==0.61.1

# Cloudinary Integration
cloudinary==1.37.0
django-cloudinary-storage==0.3.0

# Additional Supporting Packages
django-rest-framework-simplejwt==0.0.2
django==4.2.9

# Token Management
pyjwt==2.8.0

# Optional but Recommended
python-dotenv==1.0.1  # For environment variable management
pillow==10.2.0  # Image processing library